Implemented installer and windows service.
-Implemented the windows installer.
-Implamented Windows service.
-Changed db name.
-Changed db and other dependencies locations from appdata to program data.
-Refactored to fix bugs.
-Had to add file permissions to the program data folder to write to db.
-Fixed installer made it start service properly and copy data per machine.

+++ Libraries/OrbisLib/Classes/TargetManagement.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using DarkUI.Forms;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.InteropServices;
@@ -118,6 +119,12 @@ namespace OrbisSuite.Classes
 
         public bool DeleteTarget(string TargetName)
         {
+            if(TargetName.Equals(DefaultTarget.Name))
+            {
+                DarkMessageBox.ShowError($"{TargetName} is the Default Target and cant be deleted.", "Cant Delete Default Target");
+                return false;
+            }
+
             return Imports.TargetManagement.DeleteTarget(TargetName);
         }
